requirements.txt
# Dépendances de base
asyncio
edge-tts>=6.1.5
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
pdfminer.six>=20221105
pygame>=2.5.0
Pillow>=9.0.0
lxml>=4.9.0
pydub
tf-keras
tensorflow
transformers
sentence-transformers
librosa
openai-whisper
soundfile
numpy
torch

